+++ src/Files.App/App.xaml.cs
@@ -7,7 +7,6 @@ using Files.App.Contexts;
 using Files.App.Data.Models;
 using Files.App.Extensions;
 using Files.App.Filesystem;
-using Files.App.Filesystem.Cloud;
 using Files.App.Filesystem.FilesystemHistory;
 using Files.App.Helpers;
 using Files.App.ServicesImplementation;
@@ -17,6 +16,8 @@ using Files.App.Shell;
 using Files.App.Storage.FtpStorage;
 using Files.App.Storage.NativeStorage;
 using Files.App.UserControls.MultitaskingControl;
+using Files.App.Utils.Cloud;
+using Files.App.Utils.Cloud.Detector;
 using Files.App.ViewModels;
 using Files.App.ViewModels.Settings;
 using Files.App.Views;
