+++ src/Files.App/Data/Models/ItemViewModel.cs
@@ -1,13 +1,13 @@
 // Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-using Files.App.Filesystem.Cloud;
 using Files.App.Filesystem.Search;
 using Files.App.Filesystem.StorageEnumerators;
 using Files.App.Filesystem.StorageItems;
 using Files.App.Helpers.FileListCache;
 using Files.App.Shell;
 using Files.App.Storage.FtpStorage;
+using Files.App.Utils.Cloud;
 using Files.App.ViewModels.Previews;
 using Files.Backend.Helpers;
 using Files.Backend.Services;
