+++ src/Files.App/ServicesImplementation/Settings/ApplicationSettingsService.cs
@@ -1,7 +1,7 @@
 // Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-using Files.App.Serialization;
+using Files.App.Utils.Serialization;
 using Files.Backend.Services.Settings;
 
 namespace Files.App.ServicesImplementation.Settings
