+++ src/Files.App/ServicesImplementation/Settings/FoldersSettingsService.cs
@@ -1,7 +1,7 @@
 // Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-using Files.App.Serialization;
+using Files.App.Utils.Serialization;
 using Files.Shared.EventArguments;
 using Microsoft.AppCenter.Analytics;
