+++ src/Files.App/Utils/Archives/ArchiveHelpers.cs
@@ -1,42 +1,38 @@
 ﻿// Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-using CommunityToolkit.Mvvm.DependencyInjection;
 using Files.App.Dialogs;
-using Files.App.Extensions;
-using Files.App.Filesystem;
-using Files.App.Filesystem.Archive;
 using Files.App.Filesystem.StorageItems;
-using Files.App.ViewModels;
 using Files.App.ViewModels.Dialogs;
-using Files.Backend.Helpers;
-using Files.Shared.Enums;
+using Microsoft.Extensions.Logging;
 using Microsoft.UI.Xaml.Controls;
-using System;
-using System.Collections.Generic;
-using System.Diagnostics;
+using SevenZip;
 using System.IO;
-using System.Linq;
 using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
 using Windows.Storage;
 
-namespace Files.App.Helpers
+namespace Files.App.Utils.Archives
 {
+	/// <summary>
+	/// Provides static helper for archives, powered by <see cref="SevenZip"/>.
+	/// </summary>
 	public static class ArchiveHelpers
 	{
-		private static OngoingTasksViewModel OngoingTasksViewModel = Ioc.Default.GetRequiredService<OngoingTasksViewModel>();
-
-		public static bool CanDecompress(IReadOnlyList<ListedItem> selectedItems)
-		{
-			return selectedItems.Any() && selectedItems.All(x => x.IsArchive)
-				|| selectedItems.All(x => x.PrimaryItemAttribute == StorageItemTypes.File && FileExtensionHelpers.IsZipFile(x.FileExtension));
-		}
+		private static readonly OngoingTasksViewModel _ongoingTasksViewModel = Ioc.Default.GetRequiredService<OngoingTasksViewModel>();
 
 		public static bool CanCompress(IReadOnlyList<ListedItem> selectedItems)
 		{
 			return !CanDecompress(selectedItems) || selectedItems.Count > 1;
+		}
+
+		public static bool CanDecompress(IReadOnlyList<ListedItem> selectedItems)
+		{
+			return
+				selectedItems.Any() &&
+				selectedItems.All(x => x.IsArchive) ||
+				selectedItems.All(x =>
+					x.PrimaryItemAttribute == StorageItemTypes.File &&
+					FileExtensionHelpers.IsZipFile(x.FileExtension));
 		}
 
 		public static string DetermineArchiveNameFromSelection(IReadOnlyList<ListedItem> selectedItems)
@@ -44,15 +40,19 @@ namespace Files.App.Helpers
 			if (!selectedItems.Any())
 				return string.Empty;
 
-			return Path.GetFileName(
+			return
+				Path.GetFileName(
 					selectedItems.Count is 1
-					? selectedItems[0].ItemPath
-					: Path.GetDirectoryName(selectedItems[0].ItemPath
-				)) ?? string.Empty;
+						? selectedItems[0].ItemPath
+						: Path.GetDirectoryName(selectedItems[0].ItemPath))
+				?? string.Empty;
 		}
 
 		public static (string[] Sources, string directory, string fileName) GetCompressDestination(IShellPage associatedInstance)
 		{
+			if (associatedInstance.SlimContentPage.SelectedItems is null)
+				return default;
+
 			string[] sources = associatedInstance.SlimContentPage.SelectedItems
 				.Select(item => item.ItemPath)
 				.ToArray();
@@ -61,7 +61,6 @@ namespace Files.App.Helpers
 				return (sources, string.Empty, string.Empty);
 
 			string directory = associatedInstance.FilesystemViewModel.WorkingDirectory.Normalize();
-
 
 			if (App.LibraryManager.TryGetLibrary(directory, out var library) && !library.IsEmpty)
 				directory = library.DefaultSaveFolder;
@@ -76,49 +75,47 @@ namespace Files.App.Helpers
 			var archivePath = creator.GetArchivePath();
 
 			int index = 1;
-			while (File.Exists(archivePath) || System.IO.Directory.Exists(archivePath))
+
+			while (File.Exists(archivePath) || Directory.Exists(archivePath))
 				archivePath = creator.GetArchivePath($" ({++index})");
+
 			creator.ArchivePath = archivePath;
 
 			CancellationTokenSource compressionToken = new();
-			PostedStatusBanner banner = OngoingTasksViewModel.PostOperationBanner
-			(
+
+			PostedStatusBanner banner = _ongoingTasksViewModel.PostOperationBanner(
 				"CompressionInProgress".GetLocalizedResource(),
 				archivePath,
 				0,
 				ReturnResult.InProgress,
 				FileOperationType.Compressed,
-				compressionToken
-			);
+				compressionToken);
 
 			creator.Progress = banner.ProgressEventSource;
+
 			bool isSuccess = await creator.RunCreationAsync();
 
 			banner.Remove();
 
 			if (isSuccess)
 			{
-				OngoingTasksViewModel.PostBanner
-				(
+				_ongoingTasksViewModel.PostBanner(
 					"CompressionCompleted".GetLocalizedResource(),
 					string.Format("CompressionSucceded".GetLocalizedResource(), archivePath),
 					0,
 					ReturnResult.Success,
-					FileOperationType.Compressed
-				);
+					FileOperationType.Compressed);
 			}
 			else
 			{
 				NativeFileOperationsHelper.DeleteFileFromApp(archivePath);
 
-				OngoingTasksViewModel.PostBanner
-				(
+				_ongoingTasksViewModel.PostBanner(
 					"CompressionCompleted".GetLocalizedResource(),
 					string.Format("CompressionFailed".GetLocalizedResource(), archivePath),
 					0,
 					ReturnResult.Failed,
-					FileOperationType.Compressed
-				);
+					FileOperationType.Compressed);
 			}
 		}
 
@@ -129,7 +126,7 @@ namespace Files.App.Helpers
 
 			CancellationTokenSource extractCancellation = new();
 
-			PostedStatusBanner banner = OngoingTasksViewModel.PostOperationBanner(
+			PostedStatusBanner banner = _ongoingTasksViewModel.PostOperationBanner(
 				"ExtractingArchiveText".GetLocalizedResource(),
 				archive.Path,
 				0,
@@ -140,14 +137,20 @@ namespace Files.App.Helpers
 			Stopwatch sw = new();
 			sw.Start();
 
-			await FilesystemTasks.Wrap(() => ZipHelpers.ExtractArchive(archive, destinationFolder, password, banner.ProgressEventSource, extractCancellation.Token));
+			await FilesystemTasks.Wrap(() =>
+				ExtractArchive(
+					archive,
+					destinationFolder,
+					password,
+					banner.ProgressEventSource,
+					extractCancellation.Token));
 
 			sw.Stop();
 			banner.Remove();
 
 			if (sw.Elapsed.TotalSeconds >= 6)
 			{
-				OngoingTasksViewModel.PostBanner(
+				_ongoingTasksViewModel.PostBanner(
 					"ExtractingCompleteText".GetLocalizedResource(),
 					"ArchiveExtractionCompletedSuccessfullyText".GetLocalizedResource(),
 					0,
@@ -168,7 +171,7 @@ namespace Files.App.Helpers
 			if (archive is null)
 				return;
 
-			var isArchiveEncrypted = await FilesystemTasks.Wrap(() => ZipHelpers.IsArchiveEncrypted(archive));
+			var isArchiveEncrypted = await FilesystemTasks.Wrap(() => IsArchiveEncrypted(archive));
 			var password = string.Empty;
 
 			DecompressArchiveDialog decompressArchiveDialog = new();
@@ -177,6 +180,7 @@ namespace Files.App.Helpers
 				IsArchiveEncrypted = isArchiveEncrypted,
 				ShowPathSelection = true
 			};
+
 			decompressArchiveDialog.ViewModel = decompressArchiveViewModel;
 
 			ContentDialogResult option = await decompressArchiveDialog.TryShowAsync();
@@ -216,7 +220,7 @@ namespace Files.App.Helpers
 				BaseStorageFile archive = await StorageHelpers.ToStorageItem<BaseStorageFile>(selectedItem.ItemPath);
 				BaseStorageFolder currentFolder = await StorageHelpers.ToStorageItem<BaseStorageFolder>(associatedInstance.FilesystemViewModel.CurrentFolder.ItemPath);
 
-				if (await FilesystemTasks.Wrap(() => ZipHelpers.IsArchiveEncrypted(archive)))
+				if (await FilesystemTasks.Wrap(() => IsArchiveEncrypted(archive)))
 				{
 					DecompressArchiveDialog decompressArchiveDialog = new();
 					DecompressArchiveDialogViewModel decompressArchiveViewModel = new(archive)
@@ -251,7 +255,7 @@ namespace Files.App.Helpers
 				BaseStorageFolder currentFolder = await StorageHelpers.ToStorageItem<BaseStorageFolder>(associatedInstance.FilesystemViewModel.CurrentFolder.ItemPath);
 				BaseStorageFolder destinationFolder = null;
 
-				if (await FilesystemTasks.Wrap(() => ZipHelpers.IsArchiveEncrypted(archive)))
+				if (await FilesystemTasks.Wrap(() => IsArchiveEncrypted(archive)))
 				{
 					DecompressArchiveDialog decompressArchiveDialog = new();
 					DecompressArchiveDialogViewModel decompressArchiveViewModel = new(archive)
@@ -272,6 +276,123 @@ namespace Files.App.Helpers
 					destinationFolder = await FilesystemTasks.Wrap(() => currentFolder.CreateFolderAsync(Path.GetFileNameWithoutExtension(archive.Path), CreationCollisionOption.GenerateUniqueName).AsTask());
 
 				await ExtractArchive(archive, destinationFolder, password);
+			}
+		}
+
+		private static async Task<SevenZipExtractor?> GetZipFile(BaseStorageFile archive, string password = "")
+		{
+			return await FilesystemTasks.Wrap(async () =>
+			{
+				var arch = new SevenZipExtractor(await archive.OpenStreamForReadAsync(), password);
+				return arch?.ArchiveFileData is null ? null : arch; // Force load archive (1665013614u)
+			});
+		}
+
+		public static async Task<bool> IsArchiveEncrypted(BaseStorageFile archive)
+		{
+			using SevenZipExtractor? zipFile = await GetZipFile(archive);
+			if (zipFile is null)
+				return true;
+
+			return zipFile.ArchiveFileData.Any(file => file.Encrypted || file.Method.Contains("Crypto") || file.Method.Contains("AES"));
+		}
+
+		public static async Task ExtractArchive(BaseStorageFile archive, BaseStorageFolder destinationFolder, string password, IProgress<FileSystemProgress> progress, CancellationToken cancellationToken)
+		{
+			using SevenZipExtractor? zipFile = await GetZipFile(archive, password);
+			if (zipFile is null)
+				return;
+
+			var directoryEntries = new List<ArchiveFileInfo>();
+			var fileEntries = new List<ArchiveFileInfo>();
+			foreach (ArchiveFileInfo entry in zipFile.ArchiveFileData)
+			{
+				if (!entry.IsDirectory)
+					fileEntries.Add(entry);
+				else
+					directoryEntries.Add(entry);
+			}
+
+			if (cancellationToken.IsCancellationRequested) // Check if cancelled
+				return;
+
+			var directories = new List<string>();
+			try
+			{
+				directories.AddRange(directoryEntries.Select((entry) => entry.FileName));
+				directories.AddRange(fileEntries.Select((entry) => Path.GetDirectoryName(entry.FileName)));
+			}
+			catch (Exception ex)
+			{
+				App.Logger.LogWarning(ex, $"Error transforming zip names into: {destinationFolder.Path}\n" +
+					$"Directories: {string.Join(", ", directoryEntries.Select(x => x.FileName))}\n" +
+					$"Files: {string.Join(", ", fileEntries.Select(x => x.FileName))}");
+				return;
+			}
+
+			foreach (var dir in directories.Distinct().OrderBy(x => x.Length))
+			{
+				if (!NativeFileOperationsHelper.CreateDirectoryFromApp(dir, nint.Zero))
+				{
+					var dirName = destinationFolder.Path;
+					foreach (var component in dir.Split(Path.DirectorySeparatorChar, StringSplitOptions.RemoveEmptyEntries))
+					{
+						dirName = Path.Combine(dirName, component);
+						NativeFileOperationsHelper.CreateDirectoryFromApp(dirName, nint.Zero);
+					}
+				}
+
+				if (cancellationToken.IsCancellationRequested) // Check if canceled
+					return;
+			}
+
+			if (cancellationToken.IsCancellationRequested) // Check if canceled
+				return;
+
+			// Fill files
+
+			byte[] buffer = new byte[4096];
+			int entriesAmount = fileEntries.Count;
+			int entriesFinished = 0;
+			var minimumTime = new DateTime(1);
+
+			FileSystemProgress fsProgress = new(progress, true, FileSystemStatusCode.InProgress, entriesAmount);
+			fsProgress.Report();
+
+			foreach (var entry in fileEntries)
+			{
+				if (cancellationToken.IsCancellationRequested) // Check if canceled
+					return;
+
+				string filePath = Path.Combine(destinationFolder.Path, entry.FileName);
+
+				var hFile = NativeFileOperationsHelper.CreateFileForWrite(filePath);
+				if (hFile.IsInvalid)
+					return; // TODO: handle error
+
+				// We don't close hFile because FileStream.Dispose() already does that
+				using (FileStream destinationStream = new FileStream(hFile, FileAccess.Write))
+				{
+					try
+					{
+						await zipFile.ExtractFileAsync(entry.Index, destinationStream);
+					}
+					catch (Exception ex)
+					{
+						App.Logger.LogWarning(ex, $"Error extracting file: {filePath}");
+						return; // TODO: handle error
+					}
+				}
+
+				_ = new FileInfo(filePath)
+				{
+					CreationTime = entry.CreationTime > minimumTime && entry.CreationTime < entry.LastWriteTime ? entry.CreationTime : entry.LastWriteTime,
+					LastWriteTime = entry.LastWriteTime,
+				};
+
+				entriesFinished++;
+				fsProgress.ProcessedItemsCount = entriesFinished;
+				fsProgress.Report();
 			}
 		}
 	}
