+++ src/Files.App/Utils/Cloud/Detector/SynologyDriveCloudDetector.cs
@@ -6,7 +6,7 @@ using Microsoft.Data.Sqlite;
 using System.IO;
 using Windows.Storage;
 
-namespace Files.App.Filesystem.Cloud
+namespace Files.App.Utils.Cloud.Detector
 {
 	/// <summary>
 	/// Provides an utility for Synology Drive Cloud detection.
