+++ src/Files.App/Utils/Serialization/BaseObservableJsonSettings.cs
@@ -1,10 +1,9 @@
 // Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-using System.ComponentModel;
 using System.Runtime.CompilerServices;
 
-namespace Files.App.Serialization
+namespace Files.App.Utils.Serialization
 {
 	internal abstract class BaseObservableJsonSettings : BaseJsonSettings, INotifyPropertyChanged
 	{
@@ -12,7 +11,7 @@ namespace Files.App.Serialization
 
 		protected override bool Set<TValue>(TValue? value, [CallerMemberName] string propertyName = "") where TValue : default
 		{
-			if (!base.Set<TValue>(value, propertyName))
+			if (!base.Set(value, propertyName))
 				return false;
 
 			OnPropertyChanged(propertyName);
