+++ src/Files.App/Utils/Serialization/ISettingsSerializer.cs
@@ -1,7 +1,7 @@
 // Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-namespace Files.App.Serialization
+namespace Files.App.Utils.Serialization
 {
 	internal interface ISettingsSerializer
 	{
