+++ src/Files.App/Utils/Serialization/Implementation/CachingJsonSettingsDatabase.cs
@@ -1,11 +1,7 @@
 // Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-using System.Collections;
-using System.Collections.Generic;
-using System.Linq;
-
-namespace Files.App.Serialization.Implementation
+namespace Files.App.Utils.Serialization.Implementation
 {
 	internal sealed class CachingJsonSettingsDatabase : DefaultJsonSettingsDatabase
 	{
