+++ src/Files.App/Utils/Serialization/Implementation/DefaultJsonSettingsDatabase.cs
@@ -1,12 +1,9 @@
 // Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-using System;
-using System.Collections.Generic;
-using System.Diagnostics;
 using System.Text.Json;
 
-namespace Files.App.Serialization.Implementation
+namespace Files.App.Utils.Serialization.Implementation
 {
 	internal class DefaultJsonSettingsDatabase : IJsonSettingsDatabase
 	{
