+++ src/Files.App/Utils/Serialization/Implementation/DefaultSettingsSerializer.cs
@@ -1,12 +1,10 @@
 // Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-using Files.Shared.Extensions;
-using System;
 using System.IO;
 using static Files.App.Helpers.NativeFileOperationsHelper;
 
-namespace Files.App.Serialization.Implementation
+namespace Files.App.Utils.Serialization.Implementation
 {
 	internal sealed class DefaultSettingsSerializer : ISettingsSerializer
 	{
@@ -14,9 +12,9 @@ namespace Files.App.Serialization.Implementation
 
 		public bool CreateFile(string path)
 		{
-			CreateDirectoryFromApp(Path.GetDirectoryName(path), IntPtr.Zero);
+			CreateDirectoryFromApp(Path.GetDirectoryName(path), nint.Zero);
 
-			var hFile = CreateFileFromApp(path, GENERIC_READ, FILE_SHARE_READ, IntPtr.Zero, OPEN_ALWAYS, (uint)File_Attributes.BackupSemantics, IntPtr.Zero);
+			var hFile = CreateFileFromApp(path, GENERIC_READ, FILE_SHARE_READ, nint.Zero, OPEN_ALWAYS, (uint)File_Attributes.BackupSemantics, nint.Zero);
 			if (hFile.IsHandleInvalid())
 			{
 				return false;
